requirements.txt
django
django-extensions
python-decouple
djangorestframework
djangorestframework-simplejwt
drf_spectacular
drf-extensions